+++ src/main/java/org/developerden/codosseum/Application.java
@@ -1,4 +1,4 @@
-package org.developerden;
+package org.developerden.codosseum;
 
 import io.micronaut.runtime.Micronaut;
 import io.swagger.v3.oas.annotations.*;
