+++ src/test/java/org/developerden/codosseum/CodosseumTest.java
@@ -1,4 +1,4 @@
-package org.developerden;
+package org.developerden.codosseum;
 
 import io.micronaut.runtime.EmbeddedApplication;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
